refactor date component #26
