+++ src/pages/Home/ReservesRanges/ReservesRanges.tsx
@@ -4,10 +4,10 @@ import { COLORS } from '@/consts/colors.ts'
 import { IHomeChoiceFormField } from '@/pages/Home/HomeChoiceFormField/HomeChoiceFormField.tsx'
 import { HOME_OPTIONS_MAP } from '@/pages/Home/Home.consts.ts'
 import { IFormikControllers } from '../Home.module'
-import { DatePickerProps, MobileDatePicker, PickersDay } from '@mui/x-date-pickers'
 import CalendarTodayIcon from '@/icons/calendar_today'
-import dayjs from 'dayjs'
+
 import PlusIcon from '@/icons/plus_icon'
+import { DateRangePicker, DateRangePickerValueType } from './DateRangePicker'
 
 interface IReservesRanges {
   startDateProps: IHomeChoiceFormField
@@ -20,65 +20,6 @@ interface IReservesRanges {
   recruitmentTypeProps2: IHomeChoiceFormField
   recruitmentTypeProps3: IHomeChoiceFormField
 }
-
-
-type DateRangePickerValueType = {
-  start: unknown;
-  end: unknown;
-};
-
-interface DateRangePickerProps
-  extends Omit<DatePickerProps<unknown, unknown>, 'value'> {
-  value: DateRangePickerValueType | null;
-}
-
-const DateRangePicker = (props: DateRangePickerProps) => {
-  const { value, onChange, ...rest } = props;
-  const [startDate, setStartDate] = React.useState<dayjs.Dayjs | null>(null);
-  const [endDate, setEndDate] = React.useState<dayjs.Dayjs | null>(null);
-
-  return (
-
-    <MobileDatePicker
-      showToolbar={false}
-      value={value}
-      onChange={(date: any) => {
-        if (date === null) {
-          return;
-        }
-        if (startDate === null || (startDate !== null && endDate !== null)) {
-          setStartDate(date);
-          setEndDate(null);
-          return;
-        }
-        setEndDate(date);
-      }}
-      onClose={() => {
-        if (startDate === undefined || endDate === undefined || startDate === null || endDate === null) {
-          return;
-        }
-        // eslint-disable-next-line
-        onChange({ start: startDate.toDate(), end: endDate.toDate() });
-      }}
-      closeOnSelect={false}
-      renderDay={(day, _value, DayComponentProps) => {
-        //alert (typeof endDate);
-        const dayDate = day as dayjs.Dayjs;
-        const isSelected =  startDate!=null && endDate!=null && dayDate.isAfter(startDate) && dayDate.isBefore(endDate?.add(1,'days'));
-        if (isSelected) {
-          return (<PickersDay {...DayComponentProps} sx={{ backgroundColor:'#4585FD', '&:focus': {
-            backgroundColor:'#4585FD'
-          } }} />)
-        } else {
-          return (<PickersDay {...DayComponentProps} />)
-        }
-      }}
-      {...rest}
-    />
-
-  );
-};
-
 
 const ReservesRanges: React.FC<IReservesRanges> = ({ startDateProps, endDateProps, 
   startDateProps2, endDateProps2,
